relocates planning units files & minor styling
