+++ app/pages/projects/[pid]/scenarios/[sid]/edit.tsx
@@ -15,6 +15,7 @@ import { useSaveScenario, useScenario } from 'hooks/scenarios';
 import MetaIcons from 'layout/meta-icons';
 import ProjectLayout from 'layout/project';
 import Sidebar from 'layout/project/sidebar';
+import AdjustPanningUnits from 'layout/project/sidebar/scenario/grid-setup/planning-unit-status';
 import Protected from 'layout/protected';
 import ScenariosSidebarSetupFeaturesAdd from 'layout/scenarios/edit/features/set-up/add';
 import ScenariosSidebarSetupFeaturesTarget from 'layout/scenarios/edit/features/set-up/targets';
@@ -22,7 +23,6 @@ import ScenarioLock from 'layout/scenarios/edit/lock';
 import ScenarioEditMap from 'layout/scenarios/edit/map';
 import AdvancedSettings from 'layout/scenarios/edit/parameters/advanced-settings';
 import BLMCalibration from 'layout/scenarios/edit/parameters/blm-calibration';
-import AdjustPanningUnits from 'layout/scenarios/edit/planning-unit/adjust-planning-units';
 import ScenariosCostSurface from 'layout/scenarios/edit/planning-unit/cost-surface';
 import ScenariosSidebarWDPACategories from 'layout/scenarios/edit/planning-unit/protected-areas/categories';
 import ScenariosSidebarWDPAThreshold from 'layout/scenarios/edit/planning-unit/protected-areas/threshold';
